修复时间错误问题，完善数值捕获

+++ GHvacDataInfo.h
@@ -22,7 +22,7 @@ public:
 public:
     GHvacDataInfo();
     void clear();
-    QJsonObject get(int msecsSinceEpoch);
+    QJsonObject get(qint64 msecsSinceEpoch);
 
     //获取系统表
     QList<TablePtr> getSystemTables() const;
